Basic API patients endpoints support GET requests; renamed API project to REST-LIS
