--- App.config
@@ -1,63 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<configuration>
-    <configSections>
-        <sectionGroup name="applicationSettings" type="System.Configuration.ApplicationSettingsGroup, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089">
-            <section name="UniversaLIS.Properties.Settings" type="System.Configuration.ClientSettingsSection, System.Configuration.ConfigurationManager, Version=6.0.0.0, Culture=neutral, PublicKeyToken=cc7b13ffcd2ddd51" requirePermission="false" />
-            <section name="IMMULIS.Properties.Settings" type="System.Configuration.ClientSettingsSection, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" requirePermission="false" />
-        </sectionGroup>
-    </configSections>
-    <connectionStrings>
-        <add name="IMMULIS.Properties.Settings.ConnectionString" connectionString="Data Source=127.0.0.1\SQLExpress;Initial Catalog=LISDB;Integrated Security=True" />
-    </connectionStrings>
-    <startup> 
-        <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.8" />
-    </startup>
-    <applicationSettings>
-        <UniversaLIS.Properties.Settings>
-            <setting name="LIS_ID" serializeAs="String">
-                <value>LIS</value>
-            </setting>
-            <setting name="SenderAddress" serializeAs="String">
-                <value />
-            </setting>
-            <setting name="SenderPhone" serializeAs="String">
-                <value />
-            </setting>
-            <setting name="AutoSendOrders" serializeAs="String">
-                <value>False</value>
-            </setting>
-            <setting name="AutoSendInterval" serializeAs="String">
-                <value>15000</value>
-            </setting>
-        </UniversaLIS.Properties.Settings>
-        <IMMULIS.Properties.Settings>
-            <setting name="LIS_ID" serializeAs="String">
-                <value>LIS</value>
-            </setting>
-            <setting name="SenderAddress" serializeAs="String">
-                <value />
-            </setting>
-            <setting name="SenderPhone" serializeAs="String">
-                <value />
-            </setting>
-            <setting name="AutoSendOrders" serializeAs="String">
-                <value>False</value>
-            </setting>
-            <setting name="AutoSendInterval" serializeAs="String">
-                <value>15000</value>
-            </setting>
-        </IMMULIS.Properties.Settings>
-    </applicationSettings>
-  <runtime>
-    <assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
-      <dependentAssembly>
-        <assemblyIdentity name="SQLitePCLRaw.core" publicKeyToken="1488e028ca7ab535" culture="neutral" />
-        <bindingRedirect oldVersion="0.0.0.0-2.0.5.1333" newVersion="2.0.5.1333" />
-      </dependentAssembly>
-      <dependentAssembly>
-        <assemblyIdentity name="System.Runtime.CompilerServices.Unsafe" publicKeyToken="b03f5f7f11d50a3a" culture="neutral" />
-        <bindingRedirect oldVersion="0.0.0.0-5.0.0.0" newVersion="5.0.0.0" />
-      </dependentAssembly>
-    </assemblyBinding>
-  </runtime>
-</configuration>
