--- ClassDiagram1.cd
@@ -1,164 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<ClassDiagram MajorVersion="1" MinorVersion="1">
-  <Class Name="IMMULIS.ServiceMain">
-    <Position X="0.5" Y="0.5" Width="1.5" />
-    <Compartments>
-      <Compartment Name="Nested Types" Collapsed="false" />
-    </Compartments>
-    <TypeIdentifier>
-      <HashCode>AAAAABAACCACAAACEACAAAACAAAAAAAAAECAAAAAAAA=</HashCode>
-      <FileName>IMMULIService.cs</FileName>
-    </TypeIdentifier>
-  </Class>
-  <Class Name="IMMULIS.CommPort">
-    <Position X="4.25" Y="0.5" Width="1.5" />
-    <TypeIdentifier>
-      <HashCode>AAAAAAAEAAAAAAACAEAAAAAAAAAAAAAAAAAAAAAAAAA=</HashCode>
-      <FileName>CommPort.cs</FileName>
-    </TypeIdentifier>
-  </Class>
-  <Class Name="IMMULIS.CountdownTimer" Collapsed="true">
-    <Position X="8.25" Y="0.5" Width="1.5" />
-    <TypeIdentifier>
-      <HashCode>AAQAAAIIAAAIBAAAAAAAAAAAAAAAAAIAAAIAAAAAAAA=</HashCode>
-      <FileName>CountdownTimer.cs</FileName>
-    </TypeIdentifier>
-  </Class>
-  <Class Name="IMMULIS.IdleState" Collapsed="true" BaseTypeListCollapsed="true">
-    <Position X="10" Y="2" Width="1.5" />
-    <TypeIdentifier>
-      <HashCode>AAAAEAIAAAAAigAAAAAABAAAAAAAAAAAAAAAAAIAAAA=</HashCode>
-      <FileName>IdleState.cs</FileName>
-    </TypeIdentifier>
-    <Lollipop Position="0.2" Collapsed="true" />
-  </Class>
-  <Class Name="IMMULIS.LISCommState" Collapsed="true" BaseTypeListCollapsed="true">
-    <Position X="10" Y="1.25" Width="1.5" />
-    <TypeIdentifier>
-      <HashCode>AACBEAIAAgAAigAAAAAABAAIIAAAQAAICAAAAAIAAAA=</HashCode>
-      <FileName>LISCommState.cs</FileName>
-    </TypeIdentifier>
-    <Lollipop Position="0.2" Collapsed="true" />
-  </Class>
-  <Class Name="IMMULIS.Program">
-    <Position X="13.25" Y="0.5" Width="1.5" />
-    <TypeIdentifier>
-      <HashCode>AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAABAAAAAA=</HashCode>
-      <FileName>Program.cs</FileName>
-    </TypeIdentifier>
-  </Class>
-  <Class Name="IMMULIS.ProjectInstaller" Collapsed="true">
-    <Position X="13.25" Y="1.75" Width="1.5" />
-    <TypeIdentifier>
-      <HashCode>gAAAAAAAACAAAAAAAACQAAACAAAAAAAAAAAAAAAAIAA=</HashCode>
-      <FileName>ProjectInstaller.cs</FileName>
-    </TypeIdentifier>
-  </Class>
-  <Class Name="IMMULIS.RcvWaitState" Collapsed="true" BaseTypeListCollapsed="true">
-    <Position X="10" Y="2.75" Width="1.5" />
-    <TypeIdentifier>
-      <HashCode>AAAAEAIAAAAAigAAAAAABAAAAAQAAAAABAAAACAAAAA=</HashCode>
-      <FileName>RcvWaitState.cs</FileName>
-    </TypeIdentifier>
-    <Lollipop Position="0.2" Collapsed="true" />
-  </Class>
-  <Class Name="IMMULIS.ReliableSerialPort">
-    <Position X="2.5" Y="0.5" Width="1.5" />
-    <TypeIdentifier>
-      <HashCode>AAAAAAAAAAAAIAAAAAAAAAAAAAAIAhAAAAAAAAAAAAA=</HashCode>
-      <FileName>ReliableSerialPort.cs</FileName>
-    </TypeIdentifier>
-  </Class>
-  <Class Name="IMMULIS.DataReceivedArgs">
-    <Position X="6" Y="0.5" Width="1.5" />
-    <TypeIdentifier>
-      <HashCode>AAAAAAAAAAAAAAAAAAAAAAAAAAAAABAAAAAAAAAAAAA=</HashCode>
-      <FileName>ReliableSerialPort.cs</FileName>
-    </TypeIdentifier>
-  </Class>
-  <Class Name="IMMULIS.TransENQState" Collapsed="true" BaseTypeListCollapsed="true">
-    <Position X="10" Y="3.5" Width="1.5" />
-    <TypeIdentifier>
-      <HashCode>AAAAEAIAAAAAigAAAAAABAAAAAQAAAAAAAAAAAAAAAA=</HashCode>
-      <FileName>TransENQState.cs</FileName>
-    </TypeIdentifier>
-    <Lollipop Position="0.2" Collapsed="true" />
-  </Class>
-  <Class Name="IMMULIS.TransWaitState" Collapsed="true" BaseTypeListCollapsed="true">
-    <Position X="10" Y="4.25" Width="1.5" />
-    <TypeIdentifier>
-      <HashCode>AAAAEAIAAAAAigAAAAAABAAAAAQAAAAAAAAAAAAAAAA=</HashCode>
-      <FileName>TransWaitState.cs</FileName>
-    </TypeIdentifier>
-    <Lollipop Position="0.2" Collapsed="true" />
-  </Class>
-  <Class Name="IMMULIS.Properties.Resources" Collapsed="true">
-    <Position X="13.25" Y="4.5" Width="1.5" />
-    <TypeIdentifier>
-      <HashCode>AAAAAAAAAAAAAAAAAgABEAAAAQAAAAAAAAAAAAAAAIA=</HashCode>
-    </TypeIdentifier>
-  </Class>
-  <Class Name="IMMULIS.Properties.Settings" Collapsed="true">
-    <Position X="13.25" Y="3.5" Width="1.5" />
-    <TypeIdentifier>
-      <HashCode>AEACBAAAAAQAAAAAAAAAKAgAAIgBAAAAADAAAAAMAAQ=</HashCode>
-    </TypeIdentifier>
-  </Class>
-  <Class Name="IMMULIS.CommFacilitator">
-    <Position X="2.5" Y="3" Width="1.5" />
-    <TypeIdentifier>
-      <HashCode>AEAAAgKAABAAAEAFYAAgAGAAAAAEBKAAABAAgAASAAA=</HashCode>
-      <FileName>CommFacilitator.cs</FileName>
-    </TypeIdentifier>
-  </Class>
-  <Class Name="IMMULIS.Message">
-    <Position X="4.5" Y="3.25" Width="1.5" />
-    <TypeIdentifier>
-      <HashCode>gAAAAAAAAAAAQSAAAAAQAAAEAEAIAAQABhAAAEEAAAA=</HashCode>
-      <FileName>Message.cs</FileName>
-    </TypeIdentifier>
-  </Class>
-  <Class Name="IMMULIS.Constants" Collapsed="true">
-    <Position X="13.25" Y="2.75" Width="1.5" />
-    <TypeIdentifier>
-      <HashCode>AAgAAAIABAAAAABAAAAAAQAAgACAAAAAAAAAAAAAIAA=</HashCode>
-      <FileName>Constants.cs</FileName>
-    </TypeIdentifier>
-  </Class>
-  <Class Name="IMMULIS.Order">
-    <Position X="8" Y="5.75" Width="1.5" />
-    <TypeIdentifier>
-      <HashCode>AAAAABAAAAAAAAAAAAAQACAACgAAAAAAAQAAAAAAAAA=</HashCode>
-      <FileName>Order.cs</FileName>
-    </TypeIdentifier>
-  </Class>
-  <Class Name="IMMULIS.Patient">
-    <Position X="6.25" Y="5.75" Width="1.5" />
-    <TypeIdentifier>
-      <HashCode>AAEBAAAAAAAAAAAAAAAQABAAAAAAAAAAAAAAAAQAAAA=</HashCode>
-      <FileName>Patient.cs</FileName>
-    </TypeIdentifier>
-  </Class>
-  <Class Name="IMMULIS.Query">
-    <Position X="6.25" Y="3" Width="1.5" />
-    <TypeIdentifier>
-      <HashCode>AAAAAAAAAAAAAAgAAAAQAAABEAAAAAAAAAAAAAAAAAA=</HashCode>
-      <FileName>Query.cs</FileName>
-    </TypeIdentifier>
-  </Class>
-  <Class Name="IMMULIS.Result">
-    <Position X="9.75" Y="5.75" Width="1.5" />
-    <TypeIdentifier>
-      <HashCode>AAAAAAAAAAAAAAAAAAQQQAAAAAAAAACAAAAAAAAAAAA=</HashCode>
-      <FileName>Result.cs</FileName>
-    </TypeIdentifier>
-  </Class>
-  <Interface Name="IMMULIS.ILISState" Collapsed="true">
-    <Position X="10" Y="0.5" Width="1.5" />
-    <TypeIdentifier>
-      <HashCode>AAAAEAIAAAAAigAAAAAABAAAAAAAAAAAAAAAAAIAAAA=</HashCode>
-      <FileName>ILISState.cs</FileName>
-    </TypeIdentifier>
-  </Interface>
-  <Font Name="Segoe UI" Size="9" />
-</ClassDiagram>
