+++ Models/OrderRequest.cs
@@ -1,0 +1,9 @@
+﻿using System.ComponentModel.DataAnnotations.Schema;
+
+namespace UniversaLIS.Models
+{
+     [Table("OrderRequest")]
+     public class OrderRequest : Order
+     {
+     }
+}
