+++ Models/Result.cs
@@ -1,0 +1,130 @@
+﻿using System;
+using System.Collections;
+using System.Collections.Generic;
+using System.Collections.Specialized;
+using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
+using System.Text.Json;
+using System.Text.Json.Serialization;
+
+namespace UniversaLIS.Models
+{
+     public class Result
+     {
+          private OrderedDictionary elements = new OrderedDictionary();
+          private List<Comment> comments = new List<Comment>();
+          private int resultID;
+          private int orderID;
+
+          [Key]
+          public int ResultID { get => resultID; set => resultID = value; }
+          public int OrderID { get => orderID; set => orderID = value; }
+          public string? UniversalTestID { get => $"{Elements["UniversalTestID"]}"; set => Elements["UniversalTestID"] = value ?? ""; }
+          public string? ResultValue { get => $"{Elements["ResultValue"]}"; set => Elements["ResultValue"] = value ?? ""; }
+          public string? Unit { get => $"{Elements["Unit"]}"; set => Elements["Unit"] = value ?? ""; }
+          public string? RefRange { get => $"{Elements["RefRange"]}"; set => Elements["RefRange"] = value ?? ""; }
+          public string? Abnormal { get => $"{Elements["Abnormal"]}"; set => Elements["Abnormal"] = value ?? ""; }
+          public string? AbNature { get => $"{Elements["AbNature"]}"; set => Elements["AbNature"] = value ?? ""; }
+          public string? ResStatus { get => $"{Elements["ResStatus"]}"; set => Elements["ResStatus"] = value ?? ""; }
+          public string? NormsChanged { get => $"{Elements["NormsChanged"]}"; set => Elements["NormsChanged"] = value ?? ""; }
+          public string? OperatorID { get => $"{Elements["OperatorID"]}"; set => Elements["OperatorID"] = value ?? ""; }
+          public string? TestStart { get => $"{Elements["TestStart"]}"; set => Elements["TestStart"] = value ?? ""; }
+          public string? TestEnd { get => $"{Elements["TestEnd"]}"; set => Elements["TestEnd"] = value ?? ""; }
+          public string? InstrumentID { get => $"{Elements["InstrumentID"]}"; set => Elements["InstrumentID"] = value ?? ""; }
+
+          public string GetResultMessage()
+          {
+               return GetResultString();
+          }
+
+          public void SetResultMessage(string value)
+          {
+               SetResultString(value);
+          }
+
+          [NotMapped]
+          [JsonIgnore]
+          public OrderedDictionary Elements { get => elements; set => elements = value; }
+          internal List<Comment> Comments { get => comments; set => comments = value; }
+
+          private string GetResultString()
+          {
+               // Anything missing should be added as an empty string.
+               string[] elementArray = { "FrameNumber", "Sequence#", "UniversalTestID", "ResultValue", "Unit", "RefRange", "Abnormal", "AbNature", "ResStatus", "NormsChanged", "OperatorID", "TestStart", "TestEnd", "InstrumentID" };
+               foreach (var item in elementArray)
+               {
+                    if (!Elements.Contains(item))
+                    {
+                         Elements.Add(item, "");
+                    }
+               }
+               string output = Constants.STX + $"{Elements["FrameNumber"]}".Trim('R') + "R|";
+               // Concatenate the Dictionary values and return the string.
+               output += Elements["Sequence#"] + "|";
+               output += Elements["UniversalTestID"] + "|";
+               output += Elements["ResultValue"] + "|";
+               output += Elements["Unit"] + "|";
+               output += Elements["RefRange"] + "|";
+               output += Elements["Abnormal"] + "|";
+               output += Elements["AbNature"] + "|";
+               output += Elements["ResStatus"] + "|";
+               output += Elements["NormsChanged"] + "|";
+               output += Elements["OperatorID"] + "|";
+               output += Elements["TestStart"] + "|";
+               output += Elements["TestEnd"] + "|";
+               output += Elements["InstrumentID"] + Constants.CR + Constants.ETX;
+               return output;
+          }
+
+          private void SetResultString(string input)
+          {
+               string[] inArray = input.Split('|');
+               if (inArray.Length < 14)
+               {
+                    // Invalid number of elements.
+                    throw new Exception($"Invalid number of elements in result record string. Expected: 14 \tFound: {inArray.Length} \tString: \n{input}");
+               }
+               Elements["FrameNumber"] = inArray[0];
+               Elements["Sequence#"] = inArray[1];
+               Elements["UniversalTestID"] = inArray[2];
+               Elements["ResultValue"] = inArray[3];
+               Elements["Unit"] = inArray[4];
+               Elements["RefRange"] = inArray[5];
+               Elements["Abnormal"] = inArray[6];
+               Elements["AbNature"] = inArray[7];
+               Elements["ResStatus"] = inArray[8];
+               Elements["NormsChanged"] = inArray[9];
+               Elements["OperatorID"] = inArray[10];
+               Elements["TestStart"] = inArray[11];
+               Elements["TestEnd"] = inArray[12];
+               Elements["InstrumentID"] = inArray[13].Substring(0, inArray[13].IndexOf(Constants.CR));
+          }
+
+          public Result(string resultMessage)
+          {
+               SetResultString(resultMessage);
+          }
+
+          public Result() { }
+
+          public string GetJsonString()
+          {
+               OrderedDictionary fieldList = new OrderedDictionary();
+               IDictionaryEnumerator enumerator = elements.GetEnumerator();
+               while (enumerator.MoveNext())
+               {
+                    switch (enumerator.Key)
+                    {
+                         case "FrameNumber":
+                         case "Sequence#":
+                              break;
+                         default:
+                              fieldList[enumerator.Key] = enumerator.Value;
+                              break;
+                    }
+               }
+               fieldList.Add("Comments", comments);
+               return JsonSerializer.Serialize(fieldList);
+          }
+     }
+}
