+++ Properties/launchSettings.json
@@ -1,0 +1,12 @@
+{
+  "profiles": {
+    "UniversaLIS": {
+      "commandName": "Project",
+      "launchBrowser": true,
+      "environmentVariables": {
+        "ASPNETCORE_ENVIRONMENT": "Development"
+      },
+      "applicationUrl": "https://localhost:50328;http://localhost:50329"
+    }
+  }
+}
