--- appsettings.json
@@ -1,5 +1,0 @@
-{
-  "ConnectionStrings": {
-    "UniversaLIS.Properties.Settings.ConnectionString": "Data Source=127.0.0.1\\SQLExpress;Initial Catalog=LISDB;Integrated Security=True"
-  }
-}
